+++ client/src/components/myinfo/MyInfo.module.css
@@ -344,67 +344,3 @@
     opacity: 1;
   }
 }
-
-/*모달*/
-.modalContainer {
-  position: fixed;
-  background: rgba(0, 0, 0, 0.5);
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  z-index: 2;
-  animation: fadein 2s;
-  -moz-animation: fadein 2s;
-  /* Firefox */
-  -webkit-animation: fadein 2s;
-  /* Safari and Chrome */
-  -o-animation: fadein 2s;
-  /* Opera */
-}
-
-.modalBox {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  border-radius: 10px;
-  background-color: white;
-  width: 20rem;
-  height: 25rem;
-  z-index: 3;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.modalText {
-  margin-bottom: 2rem;
-  text-align: center;
-  font-size: 1.5rem;
-  color: #202020;
-}
-.btnBox {
-  width: 15rem;
-  height: 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-  margin-top: 2rem;
-}
-.modifyBtn {
-  width: 5rem;
-  height: 2.5rem;
-  padding: 0.2rem;
-  border: 1px solid gainsboro;
-  background-color: white;
-  outline: none;
-  cursor: pointer;
-  font-size: 0.8rem;
-  transition: all 0.5s;
-}
-.modifyBtn:hover {
-  background-color: #202020;
-  color: white;
-  transition: transform 0.6s cubic-bezier(0.08, 0.35, 0.13, 1.02);
-}
